Moves word emphasis related functions to utils folder
